Rename the Vc plugin to Vc_AoS to distinguish it from the Vc_SoA plugin and build the latter in the CI
