+++ CMakeLists.txt
@@ -55,7 +55,8 @@ endif()
 # Flags controlling which parts of Detray to build.
 option( DETRAY_EIGEN_PLUGIN "Build Eigen math plugin" OFF )
 option( DETRAY_SMATRIX_PLUGIN "Build ROOT/SMatrix math plugin" OFF )
-option( DETRAY_VC_PLUGIN "Build Vc based math plugin" OFF )
+option( DETRAY_VC_AOS_PLUGIN "Build Vc based AoS math plugin" OFF )
+option( DETRAY_VC_SOA_PLUGIN "Build Vc based SoA math plugin" OFF )
 option( DETRAY_SVG_DISPLAY "Build ActSVG display module" OFF)
 option( DETRAY_BUILD_SYCL "Build the SYCL sources included in detray" OFF )
 option( DETRAY_BUILD_CUDA "Build the CUDA sources included in detray"
