+++ cmake/detray-config.cmake.in
@@ -11,7 +11,8 @@
 # Detray build.
 set( DETRAY_EIGEN_PLUGIN @DETRAY_EIGEN_PLUGIN@ )
 set( DETRAY_SMATRIX_PLUGIN @DETRAY_SMATRIX_PLUGIN@ )
-set( DETRAY_VC_PLUGIN @DETRAY_VC_PLUGIN@ )
+set( DETRAY_VC_AOS_PLUGIN @DETRAY_VC_AOS_PLUGIN@ )
+set( DETRAY_VC_SOA_PLUGIN @DETRAY_VC_SOA_PLUGIN@ )
 set( DETRAY_DISPLAY @DETRAY_DISPLAY@ )
 set( DETRAY_BUILD_CUDA @DETRAY_BUILD_CUDA@ )
