+++ extern/algebra-plugins/CMakeLists.txt
@@ -35,10 +35,12 @@ FetchContent_Declare(AlgebraPlugins ${DETRAY_ALGEBRA_PLUGINS_SOURCE_FULL})
 # Options used in the build of Algebra Plugins.
 set(ALGEBRA_PLUGINS_BUILD_TESTING FALSE CACHE BOOL
    "Turn off the build of the Algebra Plugins unit tests")
+set(ALGEBRA_PLUGINS_BUILD_BENCHMARKS FALSE CACHE BOOL
+   "Turn off the build of the Algebra Plugins benchmarks")
 set(ALGEBRA_PLUGINS_INCLUDE_EIGEN ${DETRAY_EIGEN_PLUGIN} CACHE BOOL
    "Turn on the build of algebra::eigen")
-set(ALGEBRA_PLUGINS_INCLUDE_VC ${DETRAY_VC_PLUGIN} CACHE BOOL
-   "Turn on the build of algebra::vc_array")
+set(ALGEBRA_PLUGINS_INCLUDE_VC ${DETRAY_VC_AOS_PLUGIN} CACHE BOOL
+   "Turn on the build of algebra::vc_aos")
 set(ALGEBRA_PLUGINS_INCLUDE_VECMEM TRUE CACHE BOOL
    "Turn on the build of algebra::vecmem_array")
 
@@ -51,14 +53,22 @@ set(ALGEBRA_PLUGINS_SETUP_EIGEN3 ${DETRAY_EIGEN_PLUGIN}  CACHE BOOL
    "Have Algebra Plugins set up Eigen3 for itself")
 set(ALGEBRA_PLUGINS_USE_SYSTEM_EIGEN3 FALSE CACHE BOOL
    "Have Algebra Plugins pick up Eigen3 from the system")
-set(ALGEBRA_PLUGINS_SETUP_VC ${DETRAY_VC_PLUGIN} CACHE BOOL
-   "Have Algebra Plugins set up Vc for itself")
+set(ALGEBRA_PLUGINS_SETUP_VC ${DETRAY_VC_AOS_PLUGIN} OR ${DETRAY_VC_SOA_PLUGIN}
+   CACHE BOOL "Have Algebra Plugins set up Vc for itself")
 set(ALGEBRA_PLUGINS_USE_SYSTEM_VC FALSE CACHE BOOL
    "Have Algebra Plugins build Vc itself")
 set(ALGEBRA_PLUGINS_SETUP_VECMEM FALSE CACHE BOOL
    "Do not set up VecMem in Algebra Plugins")
 set(ALGEBRA_PLUGINS_SETUP_GOOGLETEST FALSE CACHE BOOL
    "Do not set up GoogleTest in Algebra Plugins")
+
+# Disable a deprecation warning coming from the Vc build.
+include( CheckCXXCompilerFlag )
+check_cxx_compiler_flag( "-Wno-deprecated-enum-enum-conversion"
+   DETRAY_HAS_NO_DEPRECATED_ENUM_ENUM_CONVERSION )
+if( DETRAY_HAS_NO_DEPRECATED_ENUM_ENUM_CONVERSION )
+   DETRAY_add_flag( CMAKE_CXX_FLAGS "-Wno-deprecated-enum-enum-conversion" )
+endif()
 
 # Get it into the current directory.
 FetchContent_MakeAvailable(AlgebraPlugins)
