+++ plugins/algebra/CMakeLists.txt
@@ -16,8 +16,8 @@ endif()
 if( DETRAY_SMATRIX_PLUGIN )
    add_subdirectory( smatrix )
 endif()
-if( DETRAY_VC_PLUGIN )
-   add_subdirectory( vc )
+if( DETRAY_VC_AOS_PLUGIN )
+   add_subdirectory( vc_aos )
 endif()
 if( DETRAY_VC_SOA_PLUGIN )
    add_subdirectory( vc_soa )
