+++ plugins/algebra/eigen/CMakeLists.txt
@@ -1,8 +1,10 @@
 # Detray library, part of the ACTS project (R&D line)
 #
-# (c) 2021-2023 CERN for the benefit of the ACTS project
+# (c) 2021-2024 CERN for the benefit of the ACTS project
 #
 # Mozilla Public License Version 2.0
+
+message( STATUS "Building the 'detray::algebra_eigen' plugin" )
 
 # A sanity check.
 if( NOT ALGEBRA_PLUGINS_INCLUDE_EIGEN )
