+++ tests/integration_tests/cpu/CMakeLists.txt
@@ -37,9 +37,9 @@ if( DETRAY_SMATRIX_PLUGIN )
    detray_add_cpu_test( smatrix )
 endif()
 
-# Build the Vc tests.
-if( DETRAY_VC_PLUGIN )
-   detray_add_cpu_test( vc )
+# Build the Vc AoS tests.
+if( DETRAY_VC_AOS_PLUGIN )
+   detray_add_cpu_test( vc_aos )
 endif()
 
 # Detector navigation tests
